Moved files and added docs to options
